Minor fixes

+++ CinemaCRMDapper/Models/Movie.cs
@@ -6,6 +6,6 @@
         public string title { get; set; }
         public int year { get; set; }
         public long budget { get; set; }
-        public short avgRate { get; set; }
+        public float avgRate { get; set; }
     }
 }
